requirements.txt
asgiref==3.3.1
astroid==2.5.1
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
DateTime==4.3
Django==2.2
django-cors-headers==3.7.0
django-csv-export-view==1.1.0
django-phonenumber-field==6.0.0
djangorestframework==3.12.4
Flask==1.1.2
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.21.4
pandas==1.3.4
phonenumberslite==8.12.37
pycparser==2.20
pylint==2.7.2
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
regex==2020.11.13
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
zope.interface==5.2.0
